pyproject.toml
[tool.poetry]
name = "repository-updater"
version = "1.3.0"
description = "Home Assistant Community Apps Repository Updater"
authors = ["Franck Nijhof <opensource@frenck.dev>"]
license = "MIT"
readme = "README.md"
packages = [{include = "repositoryupdater"}]

[tool.poetry.dependencies]
python = "^3.11"
click = "8.3.1"
crayons = "0.4.0"
emoji = "2.15.0"
GitPython = "3.1.46"
Jinja2 = "3.1.6"
PyGithub = "2.8.1"
python-dateutil = "2.9.0.post0"
PyYAML = "6.0.3"
semver = "3.0.4"

[tool.poetry.group.dev.dependencies]
zizmor = "^1.0.0"

[tool.poetry.scripts]
repository-updater = "repositoryupdater.cli:repository_updater"
repository-updater-git-askpass = "repositoryupdater.cli:git_askpass"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
